+++ python/GafferSceneTest/__init__.py
@@ -206,6 +206,7 @@ from .QuantizePrimitiveVariablesTest import QuantizePrimitiveVariablesTest
 from .RandomPrimitiveVariableTest import RandomPrimitiveVariableTest
 from .PointInstancerCoreTest import PointInstancerCoreTest
 from .PointInstancerTest import PointInstancerTest
+from .VisibilityQueryTest import VisibilityQueryTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *